`numpy.argwhere` does not work with int
A very strange feature (bug?) of numpy is that when using `numpy.argwhere`, the result can differ based on whether the sequence is a `list` or a `numpy.ndarray`. For example the results of

```
np.argwhere([0, 1, 2, 3, 4] == 4).flatten()
```

and

```
np.argwhere(np.array([0, 1, 2, 3, 4]) == 4).flatten()
```

are different. This created a lot of confusion!