Apply functioning prism to interactive and standard jobs
Sorry, this should have been multiple commits. The key takeaway is that the prism is now set as an attribute of the base lammps class, and conversions to and from Lammps coordinates are consistently applied. One tricky thing was that I had to set up the prism once at the beginning, as for the regular lammps jobs I found that two input cells which looked identical when printed gave different folding properties (probably due to some extreme sensitivity to rounding), so the back rotation matrix and unfolding became inconsistent with the cells I was actually reading.

Because the prism is now an attribute of the lammps base, the _interactive_prism got renamed _prism throughout the interactive class.

I also re-ordered the collect_dump_file method and added white space so that each property has its own code block, which improves readability.

In addition, I replaced the einstein summation calls with simple matmul  calls, which I think improves readability, but I admit this is a bit more subjective. I also replaced some dot calls with matmul in obedience to the numpy docs which explicitly say it's preferred, although my feelings there are not as strong and I haven't exhaustively replaced these uses of dot.

That about covers it. Like I said, sorry this is all one commit...