fix exception on accessing missing attribute
the previous implementation of __getattr__ tried to escalate to
object.__getattr__ if the requested attribute name is not one of the
status names, probably to keep normal attribute lookup working.  But
this is not needed because python first calls __getattribute__ to do the
normal lookup and only then calls __getattr__ if it is defined[1] and
since object doesn't define __getattr__ this gave a weird exception when
you tried to access JobStatus.status_name and got the status name wrong.

This now raises the proper AttributeError, while keeping normal
attribute access intact.

[1]: https://docs.python.org/3/reference/datamodel.html?highlight=__getattribute__#object.__getattribute__