update JOB_CLASS_DICT automatically via metaclass
`RegisterJobTypeMeta` and `JobCore.__init_subclass__` (depending on
python version) contain some magic that runs at *class creation* and
registers subtypes of JobCore in the JOB_CLASS_DICT of
`pyiron.base.job.jobtype` as they are created without the need to
manually maintain it or search all modules for subclasses as we
previously did.

`__init_subclass__` is easier to understand, it is a `@classmethod`
python calls after creating a subclass of the class it is defined on.
Its `cls` argument is then the newly created class.
https://docs.python.org/3.8/reference/datamodel.html#object.__init_subclass__

In case of python3.5 and earlier this mechanism didn't exist, so we have
to create a metaclass with the same purpose.  Normally classes are
created by python via the `type()` function/metaclass.  The `__init__`
method on a metaclass initializes the class object, the same way it
initializes the instance object when defined on a normal class.
Deriving our metaclass from type means we just have to add our logic and
then fallback on `super()`.
https://docs.python.org/3.5/library/functions.html#type