Fix LammpsControl.pressure_to_lammps()
The normalization of the "pressure" argument to LammpsControl.calc_md() was broken, in that it did not preserve "None" values, which are used to indicate directions in which the barostat should not be applied. The reason was the casting of the input list to a numpy.ndarray of type "float64", which converts None into NaN. Later code checks for the presence of None, though, and silently produces invalid commands in the LAMMPS input file.

This commit fixes this and cleans up the whole procedure. Checks for the validity of the input are all in LammpsControl.pressure_to_lammps() and are no longer repeated in calc_md() or calc_minimize().

Additionally, the handling of hydrostatic pressure (by providing a single scalar) was fixed. It previously set the shear stress components to nonzero values, too, under certain conditions. 

I believe that hydrostatic pressure works with any rotation, since the matrix representation of the corresponding stress tensor is invariant under rotation, so we can use [pressure, pressure, pressure, None, None, None] even for boxes that are rotated for LAMMPS. I implemented this here, but it might not work as I thought.