Fix #1093
The underlying problem is that `Atoms.__copy__` copies the `Visualize`
class, but does not set a the new `Atoms` object as its reference.
However since `Visualize` does not keep any state expect that reference,
we can side-step this problem by making `Atoms.visualize` a property
that returns a new `Visualize` instance on the fly.  To avoid
unnecessary instantiations `Visualize` is now a singleton.  I also went
ahead and renamed the member to `_visualize` since it is not documented
and we labeled the whole module where it comes from as internal.

This effectively reverts PR #1098