Assign local creator instance
Without this the Project.__init__ super() call instantiates the pyiron_base Creator object, which doesn't know about StructureFactory. I find it very ugly to reinstantiate the attribute here, but it's a fast and clear fix until such time as the Creator object is automatically updated with factories in a similar way that the JobFactory is updated with job classes. As a temporary patch I'll live with this.