Spinoff structures (#1199)
* Refactor structure generator methods to class

* Hide structure methods in attribute and deprecate direct access

* Migrate atom and element creation and deprecate original methods

* Patch random test

* Bind ASE's generic Atoms reading method

* Add another layer of abstraction, create, for the project interface

* Call it factory

* Synchronize with pending pyiron_base changes

* Fix deprecation warnings

* Don't deprecate yet

Let the users live in blissful ignorance until the next version increment

* Remove bleedover changes from wrong branch

* Update for pyiron_base changes, creator already exists

* Inherit from pyiron_base creator

* Make Codacy happy, but preserve in anticipation of unlocking deprecation warnings

* Better test variable name

* Remove dead code

* Remove duplicate code

* Assign local creator instance

Without this the Project.__init__ super() call instantiates the pyiron_base Creator object, which doesn't know about StructureFactory. I find it very ugly to reinstantiate the attribute here, but it's a fast and clear fix until such time as the Creator object is automatically updated with factories in a similar way that the JobFactory is updated with job classes. As a temporary patch I'll live with this.

* Rename method

* rename hkl_surface surface_hkl

Co-authored-by: Jan Janssen <jan-janssen@users.noreply.github.com>

* Revert "Remove duplicate code"

This reverts commit ee61df1022c6c6801ea21c9022e751e0f879c67a.

* Expand docstring

And circumvent codacy complaint about one-liner

* Don't encourage users to instantiate their own factories

* Enforce working pyscal version

* Maintain documentation notebook compatibility

At least until we properly deprecate the old functions and formally introduce the new Creators and Factories.

* PEP8

Co-authored-by: Jan Janssen <jan-janssen@users.noreply.github.com>